Extracting data set loading method from TestJdbcHugeStream.
